pyproject.toml
[project]
name = "payout-service"
version = "0.1.0"
description = "Сервис управления заявками на выплату"
requires-python = ">=3.12"

dependencies = [
    "Django==6.0",
    "djangorestframework==3.16.1",
    "celery==5.6.0",
    "redis==7.1.0",
    "psycopg2-binary==2.9.11",
    "django-cors-headers==4.9.0",
    "django-celery-results==2.6.0",
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-cov==7.0.0",
    "pytest-django==4.11.1",
    "pytest-mock==3.14.0",
    "ruff==0.14.9",
    "model-bakery==1.15.0"
]

[tool.uv]

[tool.pytest.ini_options]
testpaths = ["payouts/tests"]
python_files = ["tests.py", "test_*.py", "*_tests.py"]
addopts = ["-v", "--tb=short"]
DJANGO_SETTINGS_MODULE = "app.settings"
markers = [
    "postgres: тесты, требующие PostgreSQL",
    "integration: интеграционные тесты",
    "slow: медленные тесты",
]

[tool.coverage.run]
source = ["payouts"]

[tool.ruff]
target-version = "py312"
line-length = 100

extend-exclude = [
    "**/migrations/*.py",
    "**/migrations/",
    "**/migrations/**/*.py",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort (сортировка импортов)
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
    "PLC", # pylint conventions
    "PLE", # pylint errors
    "PLR", # pylint refactoring
    "PLW", # pylint warnings
]
ignore = [
    "E501",  # line too long
    "B017",  # do not assert blind exception: `Exception` (для тестов)
    "ARG001", # unused function argument (разрешаем для фикстур)
    "ARG002", # unused method argument (разрешаем для переопределения методов DRF)
    "PLR2004", # magic value used in comparison (разрешаем в тестах)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]
"tests/*.py" = ["S101", "B017", "PLR2004"]
"**/test_*.py" = ["B017", "PLR2004", "ARG001", "ARG002"]
"payouts/tests/*.py" = ["B017", "PLR2004", "ARG001", "ARG002"]
"manage.py" = ["PLC0415"]

[tool.ruff.lint.isort]
known-first-party = ["app", "payouts"]
lines-after-imports = 2
lines-between-types = 1

[tool.ruff.format]
quote-style = "double"

[tool.ruff.lint.pylint]
allow-magic-value-types = ["int", "float"]